store pointers to already fetched dirs and segments in nodes
this avoids casting them and looking for them in the map, which would
happen in `storage.Get`.  however, does this mean the values are stored
*twice*?  because if `storage.Get` just returns the *value*, not a
pointer *to* the value, then the cache would be a bit useless.  so maybe
that is a place where one might want to use `*interface{}` as a type?