use a comparable interface instead of raw integers
unfortunately, this comes at a 3x performance hit, which is quite
disappointing...

i will likely have to look into how `github.com/google/btree` and
friends do it, and maybe even consider using a specialized version
later, although i'd strongly prefer to have the btree stay generic.

another alternative might be to just pass the values as `interface{}`,
and pass a `compare` function around.  not sure if it might be faster,
but it's worth a try.