pass btset comparator as function, not as an interface method
we need to be able to compare the same type with different comparators
(for the different indexes), but we likely don't want to wrap all values
in a separate type for that.  (we should try it regardless, maybe...
the changes would be less invasive, at least.)

and yes, it's getting slower, again.  especially lookup is much slower,
likely because it has to pass around the comparator everywhere.

before:

   PASS
   BenchmarkConj	  200000	     11856 ns/op
   BenchmarkConjRandom	  200000	      9448 ns/op
   BenchmarkLookup	10000000	       133 ns/op
   ok  	_/home/lu/k/mu/collection/btset	11.603s

after:

   PASS
   BenchmarkConj	  200000	     13020 ns/op
   BenchmarkConjRandom	  200000	     10741 ns/op
   BenchmarkLookup	 3000000	       463 ns/op
   ok  	_/home/lu/k/mu/collection/btset	13.188s