change .SeekDatoms to accept datoms, not components
the full interface will be implemented by a future api package,
hopefully to be named `mu`, not `mu.api` or something like that.

the reasoning is that indexes shouldn't know about their type, and the
conversion from arbitrary input data should be done in a layer above,
not right at the "bottom", i.e. in the indexes.