stub out log tail reading in the connector
does not work yet because we need to "transact"/add the datoms in the
log the various indexes, paying attention to putting them in the correct
ones... etc.  there is an implementation of that in connection/memory,
but that results in an import cycle.  not sure yet how to resolve it.