add sketch for using the transactor
initially, and likely for a while after, there will only be an
in-process transactor.  in fact, it will be per-connection, so you will
run into problems if you try to connect to the same database using
different connections.

thinking about it, maybe we *will* write a lock to the store, but then
that would amount to given guarantees that we can't really give, because
we don't have conditional writes/puts in the stores.