we need ... a util/common package at some point
`Part`, the pattern stuff, and possibly more could live there.  likely
in multiple sub-packages, but with as few as possible dependencies.  has
the added benefit of being able to retain/specify a set "core" functions
in the api and mark the pattern stuff (and future things of that nature)
as "optional"/additional.

(one quick thought relating to that "future stuff":  we might want to do
something about schemas at some point...  but maybe not, because that's
easily done in external packages, especially because there is the
TxDatum interface that should allow niceties like that.)