work around an import cycle relating to Keyword values
absolutely not sure what we should do here.  Keyword has to be defined
in the `database` package to implement `Lookup`.  but that package
imports the `index` package, where `NewValue` is implemented.  we could
define it elsewhere, but i'm not quite sure where that would be.
*maybe* `index/value.go` should be in it's own package, so that the
`database` package does not have to import it.  not sure.  also, we may
not want to return fressian keywords anywhere, ever, but we currently
do, when returning things from the iterator.  i think that's ok, because
we only ever need our own Keyword for Lookup stuff, but not sure.

tl;dr: this is an evil hack, we need to fix it, but i'm not sure how
exactly we're going to do it, yet.