prevent `database.Keyword`s as values of `index.Value`s
these should *only* be used in the interfaces, they're invalid
elsewhere.  this means that `.Datoms` and friends *will* return standard
`fressian.Keyword`s, even though we're using the wrapped keywords
everywhere else.  however, i think that's ok because the indexes are
supposed to be low-level, and because hopefully you won't have to write
the type down too often.

remember the (new) rule: `database.Keyword`s as input,
`fressian.Keyword`s in the index and results.