erase the `Ref` type after checking it
because we don't know which type a value has when loading the index (no
attribute information loaded, no concept of db, etc), we erase the type.
allowing the comparison of refs and ints/longs would also work, but this
way seems more intentional, because it only matters in the transactor.

if we don't do this (either of both ways), the comparisons fails, and we
may get duplicate datoms.