remove the 'have a common/utils package' idea
these tend to accumulate all kinds of things, eventually containing
everything from small helpers to io utilities.  should we need it at
some point, we'll introduce tiny "helper" modules, possibly close to
their usage.

for example, we currently have code for extracting a partition from an
entity id (as an integer) in various places, that code could be moved to
the `database` package, or a separate `database/id` package.