start documenting the top-level package
for now, only the functions similar to the ones in datomic are
documented, everything else is subject to change, as we figure out how
to express these things in Go.

missing/to be improved:

- provide examples, both in the package overview, and for the individual
  functions
- find a place for "internal" documentation, about the database itself,
  concepts, philosophy, ...
- clearly indicate that the top-level package is the one that should be
  used, and that the other packages are not intended to be used in
  user-facing applications

however, there are still a lot of things missing, especially regarding
the last point.  for example, the core types are defined all over the
place, not in this package.  this means that users currently have to
import other packages to be able to use this.

there is no clear way to solving this problem, but here are a few ideas:

- provide "wrapper" types/aliases for the important types (Db, Datom,
  Index, Connection)
- define these types in the top-level packages.  (this is only possible
  if *all* the things we need for interdependencies need to be defined
  in this package.  however, this would very likely only lead to import
  cycles.)
- put everything in one (very big) package.  (we definitely don't want to
  do this.)