implement initial support for tx fns
not entirely sure about these, because the pattern is always the same.
(i.e. an outer function with the arguments to the tx fn, and an inner
function that closes over the arguments of the outer function.)

if we want to support reading *arbitrary* function calls from EDN, we
need to abstract over this, i.e. register functions (and their argument
count, maybe?) and then convert that to the registered functions.
however, we should be able to keep the current types, just the EDN side
would be a bit complicated.  (in particular, we would need to do *some*
kind of type-checking, using the `reflect` package.)