start working on tests for the indexes
these could (probably) be in the `index` package, but soon they will be
getting a bit more complex, supporting filters and other things, so they
feel more appropriate in the `database` package.

(not entirely sure about that, though, maybe we need an additional set
of more low-level tests in the `index` package?  or the reverse,
high-level tests here, and move these tests to the `index` package.)