wrap the raw index into an intermediate index
that knows which db it belongs to, so it can do filtering and other
things that depend on the db.  currently it simply returns the inner
iterator.

this is a temporary solution, it would probably be better to abandon the
`.Eavt()` methods, but that seems less idiomatic.  maybe later when we
introduce `.Datoms` on the db itself.