convert transaction ids when reading/writing them
on disk they're stored as offset from `3 * (1 << 42)` so that they use
up less space.  this is possible because when they appear in a tdata
segment or the tx value of a datom it's clear that it is a transaction
id.  however, when they appear as entity ids, they're stored as usual.