implement filtering a database
this should work with all the other options, as it's applied to the
resulting iter as the last (optional) step.

aside about go:  allowing `nil` as a valid value for a function is an
interesting choice, because it allows this neat pattern of checking for
it.  the entire concept of explicit "zero values" for types is
interesting, and allows for a few patterns like this, for example
making the zero values of mutexes and bytes.Buffers valid values that
can be used as is.  "recursive" zero values make this even more fun.

(note, however, that we choose to use explicit values for all fields,
because we want to know whether a value was set or not.  for example, we
set `since` and `asOf` to `-1` to know that they weren't set.  and we
have to initialize `attributeCache`, because you can't add values to
it's zero value.)