fix support for multiple filters
the problem was that `db` was the name of both the outer db we were building
the filter from and the argument of the closure.  this resulted in a
stack overflow, because the closure argument shadowed the outer db
binding.

we could implement this differently, possibly also be more efficient, by
storing a slice of filters and then iterating over them, building only
one closure, not one per added filter.