make asOf checking more robust
we will simply ignore negative values for `asOf`.  note that in contrast
to `since`, 0 is a valid value for `asOf`, and behaves differently as it
"goes back" to the first transaction.

we should probably filter t values smaller than zero in `db.AsOf` and
`db.Since`, maybe even return `nil` or crash/panic if that happens,
because it's clearly invalid.  alternatively, we could use it as a way
to "reset" the effects of either, because then the db wouldn't be
filtered.  (this is the current behavior, but only rather accidentally.)