Mark db.AsOf and mu.EntidAt as done
Currently, there are no separate functions for these in the top-level
package.  If we don't find a way to document their existence as
methods on a database value we will have to expose them there.  (Which
may be a good thing in itself, because then we are explicit about what
is the "public" api, and what's internal.)