Fix reading backups with logs
This means starting to use UUIDs in the data we write to disk.  We
didn't have to do this so far because we didn't write anything that had
UUIDs (except the log tail), but we will need to do this when writing
segments later as well.  To be able to do that, however, we will also
need to switch using UUIDs as values in the root and directory nodes.
(We currently convert them to strings when reading them in, but we
really should use the raw UUIDs, not strings.)