Note that porting pull and query is an option
`datascript` has implementations for both, so we could port them, just
as we did with the btset.  This would get us to feature completeness
faster, but isn't so good for future improvements, as the code would
likely contain a few clojure-isms.  Still, getting a query
implementation in the near future is a very compelling proposition.