Start porting the query implementation
Trying to port the query implementation from `datascript.query`.

In constrast to the port of `datascript.btset`, this won't "translate"
as easy.  Having ported the relational algebra and predicate parts, it's
clear that we need types and will likely only be able to copy the
general "approach", not the entire code.

This means we'll have to do a "conceptual" port first, to understand
what parts we need, then write the types for it and then implement it.