Implement a rough version of Search for eavt
There may be a problem with the way we do the matching ..., because
depending on how the iterators work, it might be that we can only
assemble the {min,max}Datom when we know which index we want to use.
Not sure yet, we'll see.

(If that's the case, given an index we can only fill in the leading
components, and we'll have to filter out the other values given the
remaining values.  If we're lucky it won't be a problem because the
iterators work "around" that.)