Implement mu.QString
It's the same as Q, but accepts a string as input.  Not sure if this is
"the right way", maybe we should introduce wrappers for these.  That
would make the API smaller, but it could have other drawbacks.  (But
given that Q already takes a string as an argument this is probably not
so bad.  For Transact we could introduce an alias for string
(TxDataEDN?) that would parse itself and then `.Resolve` the result as
usual.