Read nextT from the log correctly
Previously it would only work as long as *one* process would be running,
but not when it would be stopped in between transactions.  I.e. nextT
would be calculated correctly when doing transactions, but read wrongly
on startup.

Now, we go through the datoms from the transaction, and use the highest
T we find there.  This is probably still incorrect in case `:nextT` on
the index root is higher than the T parts on the datoms in the
transactions.  This can't occur yet because we don't write nextT to
index roots.  We will have to change this when doing proper indexing.