Fix duplicate merging for refs that were "merged"
Internally we need to compare refs as integers, because the index
doesn't really know about refs.  (That certain values are refs can only
be inferred via the schema, which is in the db itself.)  Thus, we need
to convert refs to ints before comparing them.

In this case, the bug was even trickier, because it only occurred when
temporary ids were merged with existing ones via a unique attribute.